Validación de versión
El actualizador compara la versión del repositorio y la compara con la versión local para validar si requiere actualización.

+++ prueba.bat
@@ -1,40 +1,67 @@
 @echo off
 
-REM Eliminar la carpeta "prueba2" si existe
-if exist prueba2 (
-    rmdir /s /q prueba2
-)
+REM Descargar el archivo remoto "version.txt"
+curl -o version_remote.txt https://raw.githubusercontent.com/mochos/actualizador/main/version.txt
 
-REM Verifica si la carpeta existe
-if exist "prueba" (
-    pushd "prueba"
+REM Comparar el archivo remoto con el archivo local
+fc /b "version.txt" "version_remote.txt" > nul
 
-    REM Identificar y mover los archivos que comienzan con "NB_" a una carpeta temporal
-    mkdir temp
-    for %%F in (NB_*) do (
-        move "%%F" temp\
+REM Verificar el código de error de fc (0 si son idénticos, 1 si son diferentes)
+if errorlevel 1 (
+    echo ##################################
+    echo Necesitas actualizar. Iniciando...
+    echo ##################################
+    timeout /nobreak /t 3 >nul
+    REM Aquí va el resto de tu script para la actualización
+
+    REM -----------------------------------------------
+
+    REM Eliminar la carpeta "prueba2" si existe
+    if exist prueba2 (
+        rmdir /s /q prueba2
     )
 
-    REM Eliminar todos los archivos restantes en la carpeta
-    del /q *
+    REM Verifica si la carpeta existe
+    if exist "prueba" (
+        pushd "prueba"
 
-    REM Mover de vuelta los archivos que habían comenzado con "NB_"
-    move temp\* .
-    rmdir /q /s temp
+        REM Identificar y mover los archivos que comienzan con "NB_" a una carpeta temporal
+        mkdir temp
+        for %%F in (NB_*) do (
+            move "%%F" temp\
+        )
 
-    popd
+        REM Eliminar todos los archivos restantes en la carpeta
+        del /q *
+
+        REM Mover de vuelta los archivos que habían comenzado con "NB_"
+        move temp\* .
+        rmdir /q /s temp
+
+        popd
+    )
+
+    REM Descargar el archivo ZIP del repositorio
+    curl -L -o repo.zip https://github.com/mochos/actualizador/archive/main.zip
+
+    REM Descomprimir el archivo ZIP en la carpeta temporal
+    tar -xf repo.zip --strip-components=1
+
+    REM Eliminar archivos temporales
+    del /q repo.zip
+    del /q version_remote.txt
+
+
+    REM -----------------------------------------------
+
+
+    pause
+    exit
+) else (
+    echo ########################
+    echo No necesitas actualizar.
+    echo ########################
+    del /q version_remote.txt
+    pause
+    exit
 )
-
-REM Descargar el archivo ZIP del repositorio
-curl -L -o repo.zip https://github.com/mochos/actualizador/archive/main.zip
-
-REM Descomprimir el archivo ZIP en la carpeta temporal
-tar -xf repo.zip --strip-components=1
-
-REM Eliminar archivos temporales
-del /q repo.zip
-
-
-REM Falta: Que lea el txt local y el txt del github y compare las versiones. Segun eso decide si actualizar o no.
-
-pause
